+++ src/components/mod.rs
@@ -3,7 +3,6 @@
 pub mod component;
 pub mod file_browser;
 pub mod file_preview;
-pub mod help_overlay;
 pub mod message;
 pub mod message_box;