+++ src/main.rs
@@ -1,5 +1,4 @@
 use anyhow::Result;
-mod keymap;
 use clap::Parser;
 
 // Import from library instead of declaring modules